Class Log_reader.of_file_jsonl

Read from given file as JSONL

inherit Imandrakit__.Core_classes.named

method read_events : only_above_level:Imandrakit__.Logger.level option -> filter_meta:(string * string) list -> unit -> Imandrakit__.Log_event.t Iter.t

Read events, in order, from the underlying log source.

parameter only_above_level
if Some lvl, only events more important than this level (inclusive) are returned

parameter filter_meta
a list of key/value pairs that must match