Module Imandrakit.Stats

Statistics output.

This provides a set of statistics that can be created after a solver is done, to print them all. Solvers can keep internal counters and add them to Stats.t when asked to.

type t = {

mutable stats : int Imandrakit_common.Str_map.t;

}

val to_serpack : Imandrakit_ser_pack.Ser.state -> t -> Imandrakit_ser_pack.value

val of_serpack : Imandrakit_ser_pack.Deser.state -> Imandrakit_ser_pack.value -> t

val pp : Imandrakit_common.Fmt.t -> t -> unit

val create : unit -> t

val add : t -> Imandrakit_common.Str_map.key -> int -> unit

val iter : t -> (string * int) Iter.t

Iterate on all stats