Module Util.Stopwatch

Stopwatch, to measure time intervals.

type t

val create : unit -> t

New stopwatch.

val time : t -> float

Time elapsed since creation. If the stopwatch was stopped using stop then this returns the span of time between creation and call to stop.

val stop : t -> unit

Stop the watch. It will always return the same time from now on.

val timeit : (unit -> 'a) -> float * 'a

timeit f returns how long f() took to run