Module Imandrakit_error

Universal errors.

This library provides an error type that can be used in many situations, can contain metadata (locations, etc.), has context (a bit like explicitly-managed stack traces), and can be serialized.

module Data : sig ... end

Extensible data for errors.

module Error : sig ... end

Universal error type for Imandrakit.

module Error_core : sig ... end

Universal error type for Imandrakit.

module Kind : sig ... end

Error kinds.