Module Imandrakit_error.Error_core

Universal error type for Imandrakit.

type message = {

msg : string;

data : Data.t;

bt : string option;
(*
Backtrace
*)

}

A message.

An error message is emitted at a particular place in the code. An error can contain several error messages.

val message_to_serpack : message Imandrakit_ser_pack.Ser.t

val message_of_serpack : message Imandrakit_ser_pack.Deser.t

type stack = message list

type t = {

process : string;

kind : Kind.t;

msg : message;

stack : stack;

}

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val to_serpack : t Imandrakit_ser_pack.Ser.t

val of_serpack : t Imandrakit_ser_pack.Deser.t

exception E of t

Internal error

val pp_with : show_process:bool -> t Imandrakit_common.Fmt.printer

module Message : sig ... end

val data : t -> Data.t

val get_data : 'a Data.key -> t -> 'a option

val raise_err : ?bt:Stdlib.Printexc.raw_backtrace -> t -> 'a

val add_bt : string -> t -> t

val add_ctx : message -> t -> t

val add_data : 'a Data.key -> 'a -> t -> t

val guard : ?let_pass:(exn -> bool) -> (unit -> message) -> (unit -> 'a) -> 'a

guard g f behaves like f(), excepts that if f() raises Error e, guard g f raises Error e' where e' wraps e with context error g().

parameter let_pass
if it returns true for an exception, the exception is re-raised.

type !'a result = ('a, t) result

val pp_result : (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a result -> Ppx_deriving_runtime.unit

val show_result : (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a result -> Ppx_deriving_runtime.string

val map_result : ('a -> 'b) -> 'a result -> 'b result

val iter_result : ('a -> Ppx_deriving_runtime.unit) -> 'a result -> Ppx_deriving_runtime.unit

val result_to_serpack : 'a Imandrakit_ser_pack.Ser.t -> 'a result Imandrakit_ser_pack.Ser.t

val result_of_serpack : 'a Imandrakit_ser_pack.Deser.t -> 'a result Imandrakit_ser_pack.Deser.t

val unwrap : 'a result -> 'a

unwrap e uses raise_err to unpack the result

module Infix : sig ... end

include module type of Infix

val let*! : (unit -> message) -> (unit -> 'a) -> 'a

Similar to guard