Module Imandrakit_metrics.Histogram

type t

A histogram with explicit buckets

val create : string -> buckets:float array -> t

create_int name ~buckets makes a histogram. Each item in buckets (which will be sorted) is used as the lower bound of a bucket. An additional bucket is created for values below the lowest item in buckets

val add_sample : t -> float -> unit

Add a value in the histogram.