Module Imandrakit_ser.Decode

Decoders for Ser_value.

Combinators to decode values.

module Error : sig ... end

Errors

Main combinators

type +'a t

Decode a value of type 'a

val int : int t

val bool : bool t

val string : string t

val return : 'a -> 'a t

val return_result : ('a, string) result -> 'a t

val return_result_err : ('a, Error.t) result -> 'a t

val delay : (unit -> 'a t) -> 'a t

val fail : string -> 'a t

val failf : ('a, Stdlib.Format.formatter, unit, 'b t) format4 -> 'a

val fail_err : Error.t -> 'a t

val unwrap_opt : string -> 'a option -> 'a t

Unwrap option, or fail

val any : Value.t t

val list : 'a t -> 'a list t

val tup2 : 'a t -> 'b t -> ('a * 'b) t

val tup3 : 'a t -> 'b t -> 'c t -> ('a * 'b * 'c) t

val tup4 : 'a t -> 'b t -> 'c t -> 'd t -> ('a * 'b * 'c * 'd) t

val dict_field : string -> 'a t -> 'a t

val dict_field_opt : string -> 'a t -> 'a option t

val dict_field_or : 'a -> string -> 'a t -> 'a t

val dict : Value.t Imandrakit_common.Str_map.t t

val dict_as_list : (string * Value.t) list t

val both : 'a t -> 'b t -> ('a * 'b) t

val reflect : 'a t -> Value.t -> ('a, Error.t) result t

reflect dec v returns the result of decoding v with dec

val apply : 'a t -> Value.t -> 'a t

apply dec v uses dec to decode v

val try_l : 'a t list -> 'a t

try_l fs tries each f in fs turn by turn, until one succeeds

val map_l : ('a -> 'b t) -> 'a list -> 'b list t

val fold_l : ('a -> 'b -> 'a t) -> 'a -> 'b list -> 'a t

module Infix : sig ... end

include module type of Infix

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val let+ : 'a t -> ('a -> 'b) -> 'b t

val and+ : 'a t -> 'b t -> ('a * 'b) t

val let* : 'a t -> ('a -> 'b t) -> 'b t

val and* : 'a t -> 'b t -> ('a * 'b) t

Deserializing

val run : 'a t -> Value.t -> ('a, Error.t) result

exception Fail of Error.t

val run_exn : 'a t -> Value.t -> 'a

raises Fail
in case of failure