Module Imandrakit_thread

Threading and concurrency library.

This builds on top of Moonpool but adds helpers.

module Background_thread : sig ... end

Background thread, for running tasks asynchronously in a sequential fashion.

module Executor : sig ... end

A task executor.

module Fiber : sig ... end

module Fut : sig ... end

Future

module Immlock : sig ... end

Lock with fast read path for immutable data

module LRU_cache : sig ... end

LRU cache.

module Lock : sig ... end

Value protected by a mutex

module Mvar : sig ... end

MVar.

module Once : sig ... end

Computed once. Thread safe.

module Rvar : sig ... end

Reactive variable.

module Switch : sig ... end

Switch to cancel tasks.

module Supervisor : sig ... end

Supervisor-like features

module Sync_queue : sig ... end

Blocking queue

module Thread_pool : sig ... end

Thread Pool.

module Timer : sig ... end

Timer.

Re-export

module Exn_bt = Moonpool.Exn_bt

val with_lock : Stdlib.Mutex.t -> (unit -> 'a) -> 'b