Module Imandrakit_twine.Dump

Produce a debug dump of the given Twine value

val dump_slice : ?string_ellipsis_threshold:int -> Imandrakit_bytes.Byte_slice.t -> string

Produce a debug dump of the twine blob, displaying individual values at their offset.

parameter string_ellipsis_threshold
limit above which strings start to be truncated when they're dumped

val dump_string : ?string_ellipsis_threshold:int -> string -> string

Like dump_slice